v0.32.2

+++ src/transforms/mezmo_tag_cardinality_limit/tests.rs
@@ -37,7 +37,7 @@ macro_rules! tags {
 }
 
 const fn make_transform_hashset(
-    value_limit: u32,
+    value_limit: usize,
     limit_exceeded_action: LimitExceededAction,
 ) -> TagCardinalityLimitConfig {
     TagCardinalityLimitConfig {
@@ -51,7 +51,7 @@ const fn make_transform_hashset(
 }
 
 const fn make_transform_bloom(
-    value_limit: u32,
+    value_limit: usize,
     limit_exceeded_action: LimitExceededAction,
 ) -> TagCardinalityLimitConfig {
     TagCardinalityLimitConfig {
